contract

+++ src/users/users.service.ts
@@ -8,6 +8,7 @@ import { JwtService } from '@nestjs/jwt'
 import { SignupDto } from './dto/signup.dto'
 import { LoginDto } from './dto/login.dto'
 import { UpdateDto } from './dto/update.dto'
+import { ResetPassDto } from './dto/reset-pass.dto'
 
 @Injectable()
 export class UsersService {
@@ -19,6 +20,10 @@ export class UsersService {
 
     async createUser(data: SignupDto, user: UserPayload): Promise<boolean> {
         try {
+            if ((data.role === ROLES.ADMIN || data.role === ROLES.SUPERADMIN) && user.role === ROLES.ADMIN)
+                throw new ForbiddenException('Bạn không có quyền tạo tài khoản quản trị!')
+            if (data.role === ROLES.VIEW && user.role === ROLES.SUPERADMIN)
+                throw new ForbiddenException('Bạn chỉ có quyền tạo tài khoản quản trị')
             const userExisted = await this.prisma.user.findFirst({ where: { userName: data.userName } })
             if (userExisted) throw new BadRequestException('User name is existed')
             const hash = await bcrypt.hash(data.password, 10)
@@ -26,7 +31,7 @@ export class UsersService {
                 data: {
                     userName: data.userName,
                     role: data.role,
-                    referUserId: data.role === ROLES.VIEW ? user.id : undefined,
+                    referUserId: user.id,
                     password: hash,
                     status: STATUS.ACTIVE,
                     amount: data.amount,
@@ -58,23 +63,38 @@ export class UsersService {
         }
     }
 
+    async resetPassword(data: ResetPassDto): Promise<any> {
+        try {
+            const hash = await bcrypt.hash(data.password, 10)
+            await this.prisma.user.update({
+                where: { id: data.id },
+                data: {
+                    password: hash,
+                },
+            })
+            return true
+        } catch (e) {
+            throw e
+        }
+    }
+
     async updateUser(data: UpdateDto, user: UserPayload): Promise<boolean> {
         try {
             const userUpdate = await this.prisma.user.findUnique({ where: { id: data.id } })
-            if (!userUpdate) throw new BadRequestException('User is not existed')
+            if (!userUpdate) throw new BadRequestException('Tài khoản không tồn tại')
             if (
-                (userUpdate.referUserId !== user.id && userUpdate.role === ROLES.VIEW) ||
-                (data.id !== user.id && userUpdate.role === ROLES.ADMIN)
+                (user.role === ROLES.VIEW && user.id !== userUpdate.id) ||
+                (user.role === ROLES.ADMIN &&
+                    (userUpdate.role === ROLES.SUPERADMIN ||
+                        (userUpdate.role === ROLES.ADMIN && user.id !== userUpdate.id)))
             )
-                throw new ForbiddenException('You do not have permission to update this user')
+                throw new ForbiddenException('Bạn không có quyền update user này')
             const hash = await bcrypt.hash(data.password, 10)
             const amount = data.amountAdded ? userUpdate.amount + data.amountAdded : undefined
             const totalAmount = data.amountAdded ? userUpdate.totalAmount + data.amountAdded : undefined
             await this.prisma.user.update({
                 where: { id: data.id },
                 data: {
-                    role: data.role,
-                    referUserId: data.role === ROLES.VIEW ? user.id : undefined,
                     password: hash,
                     status: STATUS.ACTIVE,
                     amount,
@@ -95,23 +115,63 @@ export class UsersService {
         }
     }
 
-    async deleteUser(deleteId: number, userId: number): Promise<any> {
+    async deleteUser(deleteId: number, user: UserPayload): Promise<any> {
         try {
-            return this.prisma.user.deleteMany({
-                where: {
-                    referUserId: userId,
-                    id: deleteId,
-                    role: ROLES.VIEW,
-                },
-            })
+            const account = await this.prisma.user.findUnique({ where: { id: deleteId } })
+            if (
+                (user.role === ROLES.SUPERADMIN && account.role === ROLES.SUPERADMIN) ||
+                (user.role === ROLES.ADMIN && (account.role === ROLES.SUPERADMIN || account.role === ROLES.ADMIN)) ||
+                user.role === ROLES.VIEW
+            )
+                throw new ForbiddenException('Bạn không có quyền xoá tài khoản này')
+            if (user.role === ROLES.SUPERADMIN) {
+                return this.prisma.user.deleteMany({
+                    where: {
+                        id: deleteId,
+                        role: { in: [ROLES.VIEW, ROLES.ADMIN] },
+                    },
+                })
+            } else if (user.role === ROLES.ADMIN) {
+                return this.prisma.user.deleteMany({
+                    where: {
+                        referUserId: user.id,
+                        id: deleteId,
+                        role: ROLES.VIEW,
+                    },
+                })
+            }
         } catch (e) {
             throw e
         }
     }
 
-    async getListUser(): Promise<any> {
+    async getListUser(user: UserPayload): Promise<any> {
         try {
-            return this.prisma.user.findMany({ where: { OR: [{ role: ROLES.VIEW }, { role: ROLES.ADMIN }] } })
+            const select = {
+                id: true,
+                userName: true,
+                referUserId: true,
+                role: true,
+                status: true,
+                amount: true,
+                totalAmount: true,
+                createdAt: true,
+            }
+            if (user.role === ROLES.SUPERADMIN)
+                return this.prisma.user.findMany({
+                    where: { OR: [{ role: ROLES.VIEW }, { role: ROLES.ADMIN }, { role: ROLES.SUPERADMIN }] },
+                    select,
+                })
+            if (user.role === ROLES.ADMIN)
+                return this.prisma.user.findMany({
+                    where: { OR: [{ referUserId: user.id }, { id: user.id }] },
+                    select,
+                })
+            if (user.role === ROLES.VIEW)
+                return this.prisma.user.findMany({
+                    where: { id: user.id },
+                    select,
+                })
         } catch (e) {
             throw e
         }
